Clear /W4 compiler warnings
The build sets /W4 (MSVC) and -Wall -Wextra -Wpedantic, but four
warnings remained:

- SimulationTime::getCurrentMonth returned uint64/uint32 division into
  a uint32 without an explicit cast (C4244).
- DistrictCommands' toupper transform lambda returned int into a char
  range (C4244); cast to char to match the CliUtils/ServiceSystem
  siblings.
- An unused groupId local in TrafficSystemTests (C4189).
- An int64/int64 ratio passed to EXPECT_NEAR in EconomySystemTests,
  converting int64 to double implicitly (C4244).

No behavior change; clean build, 138 tests pass.

+++ src/core/SimulationTime.hpp
@@ -18,7 +18,7 @@ struct SimulationTime {
   }
   
   uint32_t getCurrentMonth() const {
-    return (tickCount / ticksPerMonth);
+    return static_cast<uint32_t>(tickCount / ticksPerMonth);
   }
   
   bool isDayBoundary() const {
